made changes and used unsigned

+++ 0x00-hello_world/6-size.c
@@ -3,10 +3,10 @@
 int main(void)
 { int intType; long int longInt; long long int longlIntype; char charType; float floatType;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(intType));
-printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(longInt));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(longlIntype));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(charType));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(floatType));
+printf("Size of a char: %zu byte(s)\n",sizeof(intType));
+printf("Size of an int: %zu byte(s)\n", sizeof(longInt));
+printf("Size of a long int: %zu byte(s)\n",sizeof(longlIntype));
+printf("Size of a long long int: %zu byte(s)\n",sizeof(charType));
+printf("Size of a float: %zu byte(s)\n",sizeof(floatType));
 return (0);
 }
